backend/requirements.txt
# Django
Django>=5.0,<6.0
djangorestframework>=3.14,<4.0
django-cors-headers>=4.3,<5.0

# WebSocket
channels>=4.0,<5.0
channels-redis>=4.1,<5.0
daphne>=4.0,<5.0

# Google Drive
google-api-python-client>=2.100,<3.0
google-auth-httplib2>=0.1,<1.0
google-auth-oauthlib>=1.1,<2.0

# Utils
python-dotenv>=1.0,<2.0
Pillow>=10.0,<11.0

# Dev
pytest>=7.4,<8.0
pytest-django>=4.5,<5.0
